# Improve performance and reduce allocations in Pipeline