Improves performance and reduce allocationss in **Pipeline** by avoiding virtual method calls and using `List<T>.ForEach(Action<T> action)` instead of `ExtensionsForIEnumerableOfT.Map(this IEnumerable<T> series, Action<T> action)` extension method.

**Before:**

|                                            Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------------------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|           Activate_WithStrategies_ObjectActivated |  34.86 ns | 0.0287 ns | 0.0240 ns |      0.0076 |           - |           - |                32 B |
|        Activate_WithStrategies_ObjectNotActivated | 107.14 ns | 0.0966 ns | 0.0856 ns |      0.0323 |           - |           - |               136 B |
|        Activate_WithoutStrategies_ObjectActivated |  33.83 ns | 0.0296 ns | 0.0277 ns |      0.0076 |           - |           - |                32 B |
|     Activate_WithoutStrategies_ObjectNotActivated |  38.20 ns | 0.0277 ns | 0.0260 ns |      0.0324 |           - |           - |               136 B |
|       Deactivate_WithStrategies_ObjectDeactivated |  36.93 ns | 0.0492 ns | 0.0460 ns |      0.0076 |           - |           - |                32 B |
|    Deactivate_WithStrategies_ObjectNotDeactivated | 107.65 ns | 0.4853 ns | 0.4302 ns |      0.0323 |           - |           - |               136 B |
|    Deactivate_WithoutStrategies_ObjectDeactivated |  36.69 ns | 0.0232 ns | 0.0217 ns |      0.0076 |           - |           - |                32 B |
| Deactivate_WithoutStrategies_ObjectNotDeactivated |  38.12 ns | 0.0428 ns | 0.0401 ns |      0.0324 |           - |           - |               136 B |

**After:**

|                                            Method |     Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------------------------- |---------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|           Activate_WithStrategies_ObjectActivated | 34.80 ns | 0.0247 ns | 0.0231 ns |      0.0076 |           - |           - |                32 B |
|        Activate_WithStrategies_ObjectNotActivated | 41.44 ns | 0.0931 ns | 0.0871 ns |      0.0228 |           - |           - |                96 B |
|        Activate_WithoutStrategies_ObjectActivated | 33.76 ns | 0.0251 ns | 0.0222 ns |      0.0076 |           - |           - |                32 B |
|     Activate_WithoutStrategies_ObjectNotActivated | 27.64 ns | 0.0716 ns | 0.0634 ns |      0.0229 |           - |           - |                96 B |
|       Deactivate_WithStrategies_ObjectDeactivated | 36.89 ns | 0.0923 ns | 0.0863 ns |      0.0076 |           - |           - |                32 B |
|    Deactivate_WithStrategies_ObjectNotDeactivated | 41.75 ns | 0.1960 ns | 0.1737 ns |      0.0228 |           - |           - |                96 B |
|    Deactivate_WithoutStrategies_ObjectDeactivated | 37.18 ns | 0.0817 ns | 0.0764 ns |      0.0076 |           - |           - |                32 B |
| Deactivate_WithoutStrategies_ObjectNotDeactivated | 27.32 ns | 0.1722 ns | 0.1526 ns |      0.0229 |           - |           - |                96 B |